# Desarrollo/Proyecto Eventos/eventos-uv-api/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: EventosAPI

host: localhost:10010

# Cors config

x-google-endpoints:
  - name: 'localhost:10010'
    allowCors: 'true'

# Paths

basePath: /

schemes: [http, https]

consumes:
  - application/json
produces:
  - application/json
paths:
  
  /login:
    x-swagger-router-controller: login.controller
    
    post:
      operationId: login
      parameters:
        - name: auth
          in: body
          required: true
          schema:
            $ref: "#/definitions/Auth"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/UsuarioResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        
  /usuarios:
    x-swagger-router-controller: usuario.controller
    
    get:
      operationId: getUsuarios
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/UsuarioList"
        default:
          description: Error
          schema: 
            $ref: "#/definitions/ErrorResponse"
    
    post:
      operationId: postUsuario
      parameters:
        - name: usuario
          in: body
          required: true
          schema:
            $ref: "#/definitions/Usuario"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/UsuarioResponse"
        "409":
          description: Email Conflict
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /usuarios/{id}:
    x-swagger-router-controller: usuario.controller

    put:
      operationId: putUsuario
      parameters:
        - name: id
          type: string
          in: path
          required: true
        - name: usuario
          in: body
          schema:
            $ref: "#/definitions/Usuario"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/UsuarioResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    
    delete:
      operationId: deleteUsuario
      parameters:
        - name: id
          in: path 
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/DeleteResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
            
  /eventos:
    x-swagger-router-controller: evento.controller
    
    get:
      operationId: getEventos
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/EventoList"
        default:
          description: Error
          schema: 
            $ref: "#/definitions/ErrorResponse"
    
    post:
      operationId: postEvento
      parameters:
        - name: evento
          in: body
          required: true
          schema:
            $ref: "#/definitions/Evento"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/EventoResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /eventos/{id}:
    x-swagger-router-controller: evento.controller
  
    get:
      operationId: getEventoById
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: success
          schema:
            $ref: "#/definitions/EventoResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    put:
      operationId: putEvento
      parameters:
        - name: id
          type: string
          in: path
          required: true
        - name: evento
          in: body
          schema:
            $ref: "#/definitions/Evento"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/EventoResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    
    delete:
      operationId: deleteEvento
      parameters:
        - name: id
          in: path 
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/DeleteResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
            
  /usuario-evento:
    x-swagger-router-controller: usuario-evento.controller
    
    post:
      operationId: postUsuarioEvento
      parameters:
        - name: usuario-evento
          in: body
          required: true
          schema:
            $ref: "#/definitions/Usuario-Evento"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/Usuario-EventoResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /usuario-evento/{id}:
    x-swagger-router-controller: usuario-evento.controller
    
    delete:
      operationId: deleteUsuarioEvento
      parameters:
        - name: id
          in: path 
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/DeleteResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  
  /swagger:
    x-swagger-pipe: swagger_raw

definitions:

# Responses

  DeleteResponse:
    type: object
    properties:
      success:
        type: number
        description: Returns 1 if successful
      description:
        type: string
        description: A short comment
    required:
      - success
      - description
      
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
        
# Login
    
  Auth:
    type: object
    properties:
      email:
        type: string
      password: 
        type: string
      rol: 
        type: string
  
# Usuario

  Usuario:
    type: object
    properties:
      password: 
        type: string
      nombre: 
        type: string
      apellidos: 
        type: string
      intereses:
        type: array
        items: 
          type: string
      email: 
        type: string
      rol: 
        type: string
    required:
      - password
      - nombre
      - apellidos
      - email
      - rol
        
  UsuarioResponse:
    type: object
    properties:
      id: 
        type: string
      password: 
        type: string
      nombre: 
        type: string
      apellidos: 
        type: string
      intereses:
        type: array
        items: 
          type: string
      email: 
        type: string
      rol: 
        type: string
    required:
      - password
      - nombre
      - apellidos
      - email
      - rol
  
  UsuarioList:
    properties:
      usuarios:
        type: array
        items:
          $ref: "#/definitions/UsuarioResponse"
          
# Eventos

  Evento:
    type: object
    properties:
      nombre: 
        type: string
      descripcion:
        type: string
      invitados:
        type: string
      fecha: 
        type: string
      hora:
        type: string
      tags:
        type: array
        items:
          type: string
      foto:
        type: string
      realizado:
        type: boolean
    required:
      - nombre
      - descripcion
      - invitados
      - fecha
      - tags
      - realizado
  
  EventoResponse:
    type: object
    properties:
      id: 
        type: number
      nombre: 
        type: string
      descripcion:
        type: string
      invitados:
        type: string
      fecha: 
        type: string
      hora:
        type: string
      tags:
        type: array
        items:
          type: string
      foto:
        type: string
      realizado:
        type: boolean
    required:
      - nombre
      - descripcion
      - invitados
      - fecha
      - tags
      - realizado
      
  EventoList:
    properties:
      eventos:
        type: array
        items:
          $ref: "#/definitions/EventoResponse"  
          
# Usuario-Evento

  Usuario-Evento:
    type: object
    properties:
      usuarioId:
        type: number
      eventoId: 
        type: number
    required:
      - usuarioId
      - eventoId
      
  Usuario-EventoResponse:
    type: object
    properties:
      id:
        type: number
      usuarioId:
        type: number
      eventoId: 
        type: number
    required:
      - id
      - usuarioId
      - eventoId